backend/patient/requirements.txt
fastapi
uvicorn[standard]
httpx
PyJWT
psycopg2-binary
